backend/requirements.txt
Flask>=2.2
Flask-Cors>=3.0
requests>=2.28
Pillow>=9.0
gunicorn>=20.1
